Try to get this working for the last few weeks

+++ gogdb/model/pricerecord.py
@@ -29,13 +29,16 @@ class PriceRecord(db.Model):
     @property
     def discount(self):
         if self.price_base == 0:
+            # If the product is free the final price is 100% of the base price
             price_fract = 1
         elif self.price_final is None or self.price_base is None:
+            # No discounts for products not for sale
             return None
         else:
             price_fract = self.price_final / self.price_base
 
-        discount_rounded = round((1 - price_fract) * 100)
+        discount_rounded = int(round((1 - price_fract) * 100))
+        # Round discounts ending with 9 or 1
         if (discount_rounded % 10) == 9:
             discount_rounded += 1
         elif (discount_rounded % 10) == 1:
